Rather than assuming that the first match is the best, add a special case for java.util.List
When searching for "List", the results are java.util.List and java.awt.List, and java.util.List
is selected as the best match. This previously worked by assuming that the first
remaining match is the best, and since java.util.List is an interface, it appears earlier in
the list. It is not clear that this makes sense in general, so "List" is now handled as a
special case instead.